+++ ikii-back-microservices/ikii-back-customers/src/main/java/mx/ikii/customers/service/ICustomerDetailsServiceWrapper.java
@@ -18,6 +18,8 @@ public interface ICustomerDetailsServiceWrapper {
 	
 	CustomerDetailsResponse update(CustomerDetailsRequest customerDetailsRequest, String id);
 	
+	void toggleFavorite(String customerId, String favorite);
+	
 	void delete(String id);
 	
 }